move align logic to util class
