+++ src/main/java/frc/robot/subsystems/Swerve.java
@@ -4,7 +4,6 @@
 
 package frc.robot.subsystems;
 
-import java.util.List;
 import java.util.function.Supplier;
 
 import com.ctre.phoenix6.hardware.CANcoder;
@@ -17,16 +16,11 @@ import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.config.PIDConstants;
 import com.pathplanner.lib.config.RobotConfig;
 import com.pathplanner.lib.controllers.PPHolonomicDriveController;
-import com.pathplanner.lib.path.PathConstraints;
-
 import edu.wpi.first.math.VecBuilder;
 import edu.wpi.first.math.estimator.SwerveDrivePoseEstimator;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.geometry.Transform2d;
-import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.ChassisSpeeds;
-import edu.wpi.first.math.util.Units;
 import edu.wpi.first.util.sendable.Sendable;
 import edu.wpi.first.util.sendable.SendableBuilder;
 import edu.wpi.first.wpilibj.DriverStation;
@@ -34,15 +28,14 @@ import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
+import frc.lib.util.AlignUtil;
 import frc.lib.util.FieldUtil;
 import frc.lib.util.LimelightHelpers;
 import frc.lib.util.PoseEstimation;
 import frc.lib.util.logging.LoggedSubsystem;
 import frc.lib.util.logging.loggedObjects.LoggedField;
 import frc.lib.util.logging.loggedObjects.LoggedSweveModules;
-import frc.robot.Constants.RobotConstants;
 import frc.robot.Constants.SwerveConstants;
 import frc.robot.LoggingConstants.SwerveLogging;
 import frc.robot.Constants.VisionConstants;
@@ -52,16 +45,6 @@ public class Swerve extends SubsystemBase {
 	public SwerveDrivetrain<TalonFX, TalonFX, CANcoder> swerve = TunerConstants.swerve;
 	SwerveDrivePoseEstimator pose = new SwerveDrivePoseEstimator(swerve.getKinematics(), getRelativeYaw(), swerve.getState().ModulePositions, Pose2d.kZero);
 
-	private double coralScoreOffsetY = Units.inchesToMeters(0);
-	// private double coralScoreOffsetX = Units.inchesToMeters(32.75);
-	private double coralScoreOffsetX = 0.938 / 2;
-
-	/* offset is relative to robot */
-	private Translation2d processorOffset = new Translation2d(RobotConstants.length / 2, 0);
-
-	private Translation2d bargeOffset = new Translation2d(Units.inchesToMeters(-24 / 2.0), 0.0);
-
-	private Pose2d selectedReef = new Pose2d(0, 0, Rotation2d.kZero);
 
 	private final SwerveRequest.ApplyRobotSpeeds autoRequest = new SwerveRequest.ApplyRobotSpeeds()
 			.withDriveRequestType(DriveRequestType.Velocity);
@@ -72,8 +55,6 @@ public class Swerve extends SubsystemBase {
 	private LoggedSubsystem logger = new LoggedSubsystem("Swerve");
 	private LoggedSweveModules logged_modules;
 	private LoggedField logged_field;
-
-	private int selectedReefindex = -1;
 
 	/** Creates a new Swerve. */
 	public Swerve() {
@@ -176,100 +157,6 @@ public class Swerve extends SubsystemBase {
 		return swerve.getModule(i).getCurrentState().speedMetersPerSecond;
 	}
 
-	public Pose2d getClosestReef() {
-		double dist = 10000000;
-		int close = 0;
-		Pose2d[] reefs = FieldUtil.getReef();
-		for (int i = 0; i < reefs.length; i++) {
-			Pose2d reef = reefs[i];
-			double distToReef = PoseEstimation.getEstimatedPose().getTranslation().getDistance(reef.getTranslation());
-			if (distToReef < dist) {
-				dist = distToReef;
-				close = i;
-			}
-		}
-
-		return reefs[close];
-	}
-
-	public Pose2d getReefScoreSpot(Pose2d reef) {
-		Translation2d tmp = new Translation2d(coralScoreOffsetX, coralScoreOffsetY);
-		tmp.rotateBy(reef.getRotation());
-
-		reef = reef.plus(new Transform2d(tmp.getX(), tmp.getY(), Rotation2d.k180deg));
-		return reef;
-	}
-
-	public void selectReef(int i) {
-		selectedReef = FieldUtil.getReef()[i];
-		selectedReefindex = i;
-	}
-
-	public Pose2d getSelectedReef() {
-		return selectedReef;
-	}
-
-	public Command driveToPose(Pose2d target) {
-		PathConstraints constraints = new PathConstraints(SwerveConstants.maxspeed, 4.0,
-				SwerveConstants.maxturn, Units.degreesToRadians(720));
-		return AutoBuilder.pathfindToPose(target, constraints, 0.0);
-	}
-
-	public Command driveToClosestReef() {
-		return driveToPose(getReefScoreSpot(getClosestReef()));
-	}
-
-	public Command driveToSelectedReef() {
-		return driveToSelectedReef(selectedReefindex);
-	}
-
-	public Command driveToSelectedReef(int i) {
-		if (i < 0 || i > 5)
-			return Commands.none();
-		selectReef(i);
-		if (selectedReef.equals(new Pose2d(0, 0, Rotation2d.kZero)))
-			return Commands.none();
-
-		return driveToPose(getReefScoreSpot(selectedReef));
-	}
-
-	public Command driveToLeftCoralStation() {
-		return driveToPose(FieldUtil.getLeftCoralStation());
-	}
-
-	public Command driveToRightCoralStation() {
-		return driveToPose(FieldUtil.getRightCoralStation());
-	}
-
-	public Pose2d getClosestBarge() {
-		Pose2d target = PoseEstimation.getEstimatedPose()
-				.nearest(List.of(FieldUtil.getAllianceSideBargeCenter(), FieldUtil.getOpponateSideBargeCenter()));
-		return target.plus(new Transform2d(bargeOffset, Rotation2d.kZero));
-	}
-
-	public Pose2d getClosestCoralStations() {
-		Pose2d target = PoseEstimation.getEstimatedPose()
-				.nearest(List.of(FieldUtil.getLeftCoralStation(), FieldUtil.getRightCoralStation()));
-		return target;
-	}
-
-	public Command driveToClosestCoralStation() {
-		return driveToPose(getClosestCoralStations());
-	}
-
-	public Command driveToClosestBarge() {
-		return driveToPose(getClosestBarge());
-	}
-
-	public Command driveToProcessor() {
-		Translation2d tmp = processorOffset;
-		tmp.rotateBy(FieldUtil.getProcessor().getRotation());
-
-		Pose2d processorScorePos = FieldUtil.getProcessor().plus(new Transform2d(tmp.getX(), tmp.getY(), Rotation2d.k180deg));
-
-		return driveToPose(processorScorePos);
-	}
-
 	@Override
 	public void periodic() {
 			updateVision();
@@ -317,9 +204,9 @@ public class Swerve extends SubsystemBase {
 		logged_modules = new LoggedSweveModules("modules", logger, this, SwerveLogging.Modules);
 
 		logged_field.addPose2d("PoseEstimation", () -> PoseEstimation.getEstimatedPose(), true);
-		logged_field.addPose2d("Closest Reef", this::getClosestReef, true);
-		logged_field.addPose2d("Selected Reef", this::getSelectedReef, true);
-		logged_field.addPose2d("Closest Barge", this::getClosestBarge, true);
+		logged_field.addPose2d("Closest Reef", () -> AlignUtil.getClosestReef(), true);
+		logged_field.addPose2d("Selected Reef", () -> AlignUtil.getSelectedReef(), true);
+		logged_field.addPose2d("Closest Barge", () -> AlignUtil.getClosestBarge(), true);
 		logger.add(logged_field);
 
 		logger.add(logged_modules);
